fix: form test setup change
